pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "newsapi-python>=0.2.7",
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "scikit-learn>=1.5.2",
    "streamlit>=1.39.0",
    "yfinance>=0.2.44",
]
